Update install.sh
If the container-service plugin is already installed. The command "bx plugin install container-service -r Bluemix" asks for 
"Plug-in 'container-service 0.1.397' was already installed. Do you want to update it with 'container-service 0.1.401' or not? (yes/no)> yes". This causes the deploy to fail. Changing the command to "bx plugin install container-service -r Bluemix -f", by forcing the install solves this issue.